refactor(test): renamed test base folder to shared
