+++ test/shared/npm.ts
@@ -17,9 +17,9 @@ export function testPackageJsonFiles(snapshot: SynthOutput): void {
 export function testGitAttributes(snapshot: SynthOutput): void {
   const patterns: RegExp[] = [
     /\/\.npmignore linguist-generated( $|\s|$)/m,
-    /\/\.eslintrc\.json linguist-generated( $|\s|$)/m,
+    // /\/\.eslintrc\.json linguist-generated( $|\s|$)/m,
     /\/tsconfig\.dev\.json linguist-generated( $|\s|$)/m,
-    /\/tsconfig\.json linguist-generated( $|\s|$)/m,
+    // /\/tsconfig\.json linguist-generated( $|\s|$)/m,
   ];
 
   common.testGitAttributes(snapshot, patterns);
