+++ test/shared/samplecode.ts
